requirements.txt
fastapi
uvicorn[standard]
transformers>=4.30
torch
pillow
python-multipart